Refactored code base into separate files
