+++ projects/01_fyyur/starter_code/application/static/fonts/fontawesome-webfont.ttf
