+++ projects/01_fyyur/starter_code/application/static/img/front-splash.jpg
