+++ projects/01_fyyur/starter_code/application/static/js/script.js
