+++ projects/01_fyyur/starter_code/application/templates/pages/artists.html
