+++ projects/01_fyyur/starter_code/requirements.txt
@@ -1,4 +1,21 @@
-babel
+alembic==1.7.3
+Babel==2.9.1
+click==8.0.1
+Flask==2.0.1
+Flask-Migrate==3.1.0
+Flask-Moment==1.0.2
+Flask-SQLAlchemy==2.5.1
+Flask-WTF==0.15.1
+greenlet==1.1.2
+importlib-resources==5.2.2
+itsdangerous==2.0.1
+Jinja2==3.0.1
+Mako==1.1.5
+MarkupSafe==2.0.1
 python-dateutil==2.6.0
-flask-moment
-flask-wtf
+pytz==2021.1
+six==1.16.0
+SQLAlchemy==1.4.25
+Werkzeug==2.0.1
+WTForms==2.3.3
+zipp==3.6.0
