+++ projects/01_fyyur/starter_code/run.py
@@ -1,0 +1,3 @@
+from application import app
+app.env="development"
+app.run(debug=True)
